+++ packages/react-querybuilder/src/utils/index.ts
@@ -10,5 +10,6 @@ export * from './optGroupUtils';
 export * from './parseSQL';
 export * from './pathUtils';
 export * from './prepareQueryObjects';
+export * from './queryTools';
 export * from './regenerateIDs';
 export * from './uniq';
